+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Changes between versions
 ### v0.6.5
 
 * throw exception if neither column nor function are set in orderBy
-* construct the order by statement when the select statement is being built
+* construct the group by and order by statements when the select statement is being
+built, since the table may not yet have been set to the query builder
 
 ### v0.6.4